more std::string -> char* conversions in Random

+++ pysrc/Random.cpp
@@ -133,7 +133,7 @@ namespace galsim {
         static UniformDeviate* construct(const bp::object& seed) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct UniformDeviate from given seed.");
-            return new UniformDeviate(0);
+            return new UniformDeviate( (long)0);
         }
 
         static void wrap() {
@@ -158,7 +158,7 @@ namespace galsim {
         static GaussianDeviate* construct(const bp::object& seed, double mean, double sigma) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct GaussianDeviate from given seed.");
-            return new GaussianDeviate(0, mean, sigma);
+            return new GaussianDeviate( (long)0, mean, sigma);
         }
 
         static void wrap() {
@@ -194,7 +194,7 @@ namespace galsim {
         static BinomialDeviate* construct(const bp::object& seed, int N, double p) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct BinomialDeviate from given seed.");
-            return new BinomialDeviate(0, N, p);
+            return new BinomialDeviate( (long)0, N, p);
         }
 
         static void wrap() {
@@ -230,7 +230,7 @@ namespace galsim {
         static PoissonDeviate* construct(const bp::object& seed, double mean) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct PoissonDeviate from given seed.");
-            return new PoissonDeviate(0, mean);
+            return new PoissonDeviate( (long)0, mean);
         }
 
         static void wrap() {
@@ -264,7 +264,7 @@ namespace galsim {
         static WeibullDeviate* construct(const bp::object& seed, double a, double b) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct WeibullDeviate from given seed.");
-            return new WeibullDeviate(0, a, b);
+            return new WeibullDeviate( (long)0, a, b);
         }
 
         static void wrap() {
@@ -301,7 +301,7 @@ namespace galsim {
         static GammaDeviate* construct(const bp::object& seed, double k, double theta) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct GammaDeviate from given seed.");
-            return new GammaDeviate(0, k, theta);
+            return new GammaDeviate( (long)0, k, theta);
         }
 
         static void wrap() {
@@ -337,7 +337,7 @@ namespace galsim {
         static Chi2Deviate* construct(const bp::object& seed, double n) {
             if (seed.ptr() != Py_None)
                 throw std::runtime_error("Cannot construct Chi2Deviate from given seed.");
-            return new Chi2Deviate(0, n);
+            return new Chi2Deviate( (long)0, n);
         }
 
         static void wrap() {
